Add rts::Frame with serialization and deserialization.
rts::Frame is an abstraction of the "base frame" in the Somfy RTS
protocol as shown in patent US8189620B2. Specifically, it's the frame
described by the "old protocol" mentioned in the patent. In the "old
protocol", the key field is a simple counter, and the rolling code
function is 'add one'.

The details of this implementation were determined by experimenting with
a Telis 4 RTS remote and a HackRF.