Add rts::TransmitFrame and TransmitInterface.
TransmitFrame() generates the pulses and timing that are expected for
the RTS protocol. TransmitInterface is an abstract interface for
connecting TransmitFrame() to a real RF transmitter in a
platform-agnostic way. For example, clients on Arduino would likely
implement TransmitInterface using digitalWrite() and
delayMicroseconds().

A simple fake TransmitInterface implementation and a test for
TransmitFrame are also added here.